Format test cases

+++ src/extension/commands/delete-test-case.command.ts
@@ -17,18 +17,12 @@ const callback = (context: ExtensionContext): any => {
     const userInput = await window.showInputBox({
       prompt: "Type 'delete all' to confirm.",
       // placeHolder: "Type 'delete all' to proceed",
-      validateInput: (text) => {
-        return text === "delete" ? null : "Please type 'delete' to confirm";
-      },
+      validateInput: (text) => (text === "delete" ? null : "Please type 'delete' to confirm"),
     });
-
     if (userInput !== "delete") return;
 
     const editor = window.activeTextEditor;
-
-    if (!editor) {
-      throw new Error("VS Code editor not found");
-    }
+    if (!editor) throw new Error("VS Code editor not found");
 
     const line = CommandUtils.getEditorLine(editor);
     const panel = CommandUtils.createWebviewPanel(context, command, "Delete test case");
@@ -36,18 +30,10 @@ const callback = (context: ExtensionContext): any => {
     const handleReceiveMessage = async (message: Message) => {
       try {
         const settings = SettingsUtils.getSettings(uri.fsPath);
-        if (!settings) {
-          throw ErrorUtils.createSettingsError();
-        }
+        if (!settings) throw ErrorUtils.createSettingsError();
 
         const description = FileUtils.extractTestCasesDescriptions(line).pop();
-        if (!description || !description.title) {
-          throw new Error("Test case description not found");
-        }
-
-        if (!description.id) {
-          throw new Error("Cannot delete test case without an ID");
-        }
+        if (!description || !description.title || !description.id) throw new Error("Test case description not found");
 
         // Delete in TestRail
         const testCase = await TestRailService.deleteTestCase(settings.project, description.id);
@@ -55,7 +41,7 @@ const callback = (context: ExtensionContext): any => {
 
         // Update file content
         const fileContent = FileUtils.getFileContent(uri.fsPath);
-        fileContent.replace(`${testCase.id}: ${testCase.title}`, testCase.title || "");
+        fileContent.replace(`${testCase.id}: ${testCase.title}`, `${testCase.title}`);
         FileUtils.saveFileContent(uri.fsPath, fileContent);
 
         window.showInformationMessage("Test case deleted!");
